requirements.txt
asgiref==3.6.0
Django==4.0.10
djangorestframework==3.13.1
gunicorn==20.1.0
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
whitenoise==6.0.0
